+++ src/FunFair.BuildCheck.ProjectChecks/ReferencedPackages/UsingXUnitMustIncludeCoverletMsBuild.cs
@@ -4,17 +4,17 @@ using Microsoft.Extensions.Logging;
 namespace FunFair.BuildCheck.ProjectChecks.ReferencedPackages
 {
     /// <summary>
-    ///     Checks that the Test SDK is installed for test projects.
+    ///     Checks that the Coverlet Ms Build is installed for test projects.
     /// </summary>
     [SuppressMessage(category: "ReSharper", checkId: "ClassNeverInstantiated.Global", Justification = "Created by DI")]
-    public sealed class UsingXUnitMustIncludeVisualStudioTestPlatform : HasAppropriateNonAnalysisPackages
+    public sealed class UsingXUnitMustIncludeCoverletMsBuild : HasAppropriateNonAnalysisPackages
     {
         /// <summary>
         ///     Constructor.
         /// </summary>
         /// <param name="logger">Logging.</param>
-        public UsingXUnitMustIncludeVisualStudioTestPlatform(ILogger<UsingXUnitMustIncludeVisualStudioTestPlatform> logger)
-            : base(detectPackageId: @"xunit", mustIncludePackageId: @"Microsoft.NET.Test.Sdk", logger: logger)
+        public UsingXUnitMustIncludeCoverletMsBuild(ILogger<UsingXUnitMustIncludeCoverletMsBuild> logger)
+            : base(detectPackageId: @"xunit", mustIncludePackageId: @"coverlet.msbuild", logger: logger)
         {
         }
     }